Joomla upgraded to 2.5.1.

+++ administrator/components/com_templates/models/style.php
@@ -1,9 +1,8 @@
 <?php
 /**
- * @version		$Id: style.php 21032 2011-03-29 16:38:31Z dextercowley $
  * @package		Joomla.Administrator
  * @subpackage	com_templates
- * @copyright	Copyright (C) 2005 - 2011 Open Source Matters, Inc. All rights reserved.
+ * @copyright	Copyright (C) 2005 - 2012 Open Source Matters, Inc. All rights reserved.
  * @license		GNU General Public License version 2 or later; see LICENSE.txt
  */
 
@@ -80,6 +79,11 @@ class TemplatesModelStyle extends JModelAdmin
 				if (!$user->authorise('core.delete', 'com_templates')) {
 					throw new Exception(JText::_('JERROR_CORE_DELETE_NOT_PERMITTED'));
 				}
+				// You should not delete a default style
+				if ($table->home != '0'){
+					JError::raiseWarning(SOME_ERROR_NUMBER, Jtext::_('COM_TEMPLATES_STYLE_CANNOT_DELETE_DEFAULT_STYLE'));
+					return false;
+				}
 
 				if (!$table->delete($pk)) {
 					$this->setError($table->getError());
@@ -91,7 +95,7 @@ class TemplatesModelStyle extends JModelAdmin
 				return false;
 			}
 		}
-		
+
 		// Clean cache
 		$this->cleanCache();
 
@@ -130,12 +134,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 				// Alter the title.
 				$m = null;
-				if (preg_match('#\((\d+)\)$#', $table->title, $m)) {
-					$table->title = preg_replace('#\(\d+\)$#', '('.($m[1] + 1).')', $table->title);
-				}
-				else {
-					$table->title .= ' (2)';
-				}
+				$table->title = $this->generateNewTitle(null, null, $table->title);
 
 				if (!$table->check() || !$table->store()) {
 					throw new Exception($table->getError());
@@ -145,11 +144,33 @@ class TemplatesModelStyle extends JModelAdmin
 				throw new Exception($table->getError());
 			}
 		}
-		
+
 		// Clean cache
 		$this->cleanCache();
 
 		return true;
+	}
+
+	/**
+	 * Method to change the title.
+	 *
+	 * @param   integer  $category_id  The id of the category.
+	 * @param   string   $alias        The alias.
+	 * @param   string   $title        The title.
+	 *
+	 * @return	string  New title.
+	 * @since	1.7.1
+	 */
+	protected function generateNewTitle($category_id, $alias, $title)
+	{
+		// Alter the title
+		$table = $this->getTable();
+		while ($table->load(array('title'=>$title)))
+		{
+			$title = JString::increment($title);
+		}
+
+		return $title;
 	}
 
 	/**
@@ -250,7 +271,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 			// Convert the params field to an array.
 			$registry = new JRegistry;
-			$registry->loadJSON($table->params);
+			$registry->loadString($table->params);
 			$this->_cache[$pk]->params = $registry->toArray();
 
 			// Get the template XML.
@@ -287,7 +308,7 @@ class TemplatesModelStyle extends JModelAdmin
 	 * @throws	Exception if there is an error in the form event.
 	 * @since	1.6
 	 */
-	protected function preprocessForm(JForm $form, $data, $group = '')
+	protected function preprocessForm(JForm $form, $data, $group = 'content')
 	{
 		// Initialise variables.
 		$clientId	= $this->getState('item.client_id');
@@ -318,9 +339,9 @@ class TemplatesModelStyle extends JModelAdmin
 
 		// Disable home field if it is default style
 
-		if ((is_array($data) && array_key_exists('home',$data) && $data['home']=='1')
+		if ((is_array($data) && array_key_exists('home', $data) && $data['home']=='1')
 			|| ((is_object($data) && isset($data->home) && $data->home=='1'))){
-			$form->setFieldAttribute('home','readonly','true');
+			$form->setFieldAttribute('home', 'readonly', 'true');
 		}
 
 		// Attempt to load the xml file.
@@ -350,6 +371,13 @@ class TemplatesModelStyle extends JModelAdmin
 	 */
 	public function save($data)
 	{
+		// Detect disabled extension
+		$extension = JTable::getInstance('Extension');
+		if ($extension->load(array('enabled' => 0, 'type' => 'template', 'element' => $data['template'], 'client_id' => $data['client_id']))) {
+			$this->setError(JText::_('COM_TEMPLATES_ERROR_SAVE_DISABLED_TEMPLATE'));
+			return false;
+		}
+
 		// Initialise variables;
 		$dispatcher = JDispatcher::getInstance();
 		$table		= $this->getTable();
@@ -365,11 +393,11 @@ class TemplatesModelStyle extends JModelAdmin
 			$isNew = false;
 		}
 		if (JRequest::getVar('task') == 'save2copy') {
-		$data['title'] .= ' '.JText::_('JGLOBAL_COPY');
-		$data['home'] = 0;
-		$data['assigned'] ='';
+			$data['title'] = $this->generateNewTitle(null, null, $data['title']);
+			$data['home'] = 0;
+			$data['assigned'] ='';
 		}
-		
+
 		// Bind the data.
 		if (!$table->bind($data)) {
 			$this->setError($table->getError());
@@ -399,7 +427,7 @@ class TemplatesModelStyle extends JModelAdmin
 		}
 
 		$user = JFactory::getUser();
-		if ($user->authorise('core.edit','com_menus') && $table->client_id==0) {
+		if ($user->authorise('core.edit', 'com_menus') && $table->client_id==0) {
 			$n		= 0;
 			$db		= JFactory::getDbo();
 			$user	= JFactory::getUser();
@@ -436,7 +464,7 @@ class TemplatesModelStyle extends JModelAdmin
 			$n += $db->getAffectedRows();
 			if ($n > 0) {
 				$app = JFactory::getApplication();
-				$app->enQueueMessage(JText::plural('COM_TEMPLATES_MENU_CHANGED',$n));
+				$app->enQueueMessage(JText::plural('COM_TEMPLATES_MENU_CHANGED', $n));
 			}
 		}
 
@@ -470,26 +498,23 @@ class TemplatesModelStyle extends JModelAdmin
 			throw new Exception(JText::_('JLIB_APPLICATION_ERROR_EDITSTATE_NOT_PERMITTED'));
 		}
 
-		// Lookup the client_id.
-		$db->setQuery(
-			'SELECT client_id' .
-			' FROM #__template_styles' .
-			' WHERE id = '.(int) $id
-		);
-		$clientId = $db->loadResult();
-
-		if ($error = $db->getErrorMsg()) {
-			throw new Exception($error);
-		}
-		else if (!is_numeric($clientId)) {
+		$style = JTable::getInstance('Style', 'TemplatesTable');
+		if (!$style->load((int)$id)) {
 			throw new Exception(JText::_('COM_TEMPLATES_ERROR_STYLE_NOT_FOUND'));
 		}
+
+		// Detect disabled extension
+		$extension = JTable::getInstance('Extension');
+		if ($extension->load(array('enabled' => 0, 'type' => 'template', 'element' => $style->template, 'client_id' => $style->client_id))) {
+			throw new Exception(JText::_('COM_TEMPLATES_ERROR_SAVE_DISABLED_TEMPLATE'));
+		}
+
 
 		// Reset the home fields for the client_id.
 		$db->setQuery(
 			'UPDATE #__template_styles' .
 			' SET home = \'0\'' .
-			' WHERE client_id = '.(int) $clientId .
+			' WHERE client_id = '.(int) $style->client_id .
 			' AND home = \'1\''
 		);
 
@@ -510,7 +535,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 		// Clean the cache.
 		$this->cleanCache();
-		
+
 		return true;
 	}
 
@@ -544,10 +569,10 @@ class TemplatesModelStyle extends JModelAdmin
 		if ($error = $db->getErrorMsg()) {
 			throw new Exception($error);
 		}
-		else if (!is_numeric($style->client_id)) {
+		elseif (!is_numeric($style->client_id)) {
 			throw new Exception(JText::_('COM_TEMPLATES_ERROR_STYLE_NOT_FOUND'));
 		}
-		else if ($style->home=='1') {
+		elseif ($style->home=='1') {
 			throw new Exception(JText::_('COM_TEMPLATES_ERROR_CANNOT_UNSET_DEFAULT_STYLE'));
 		}
 
@@ -564,7 +589,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 		// Clean the cache.
 		$this->cleanCache();
-		
+
 		return true;
 	}
 
@@ -578,13 +603,14 @@ class TemplatesModelStyle extends JModelAdmin
 	{
 		return (object) array('key' => $this->helpKey, 'url' => $this->helpURL);
 	}
-	
+
 	/**
 	 * Custom clean cache method
 	 *
 	 * @since	1.6
 	 */
-	function cleanCache() {
+	protected function cleanCache($group = null, $client_id = 0)
+	{
 		parent::cleanCache('com_templates');
 		parent::cleanCache('_system');
 	}
